+++ affirm/src/main/java/com/affirm/android/AffirmPromotionButton.java
@@ -17,30 +17,30 @@ import java.util.Locale;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatButton;
 
 import static com.affirm.android.AffirmColor.AFFIRM_COLOR_TYPE_WHITE;
-import static com.affirm.android.AffirmLogoType.AFFIRM_DISPLAY_TYPE_TEXT;
 import static com.affirm.android.AffirmConstants.LOGO_PLACEHOLDER;
+import static com.affirm.android.AffirmLogoType.AFFIRM_DISPLAY_TYPE_TEXT;
 
-public class AffirmPromotionLabel extends AppCompatTextView {
+public class AffirmPromotionButton extends AppCompatButton {
 
     private Paint mPaint;
     private AffirmLogoType mAffirmLogoType;
     private AffirmColor mAffirmColor;
 
-    public AffirmPromotionLabel(@NonNull Context context) {
+    public AffirmPromotionButton(@NonNull Context context) {
         this(context, null);
     }
 
-    public AffirmPromotionLabel(@NonNull Context context,
-                                @Nullable AttributeSet attrs) {
+    public AffirmPromotionButton(@NonNull Context context,
+                                 @Nullable AttributeSet attrs) {
         this(context, attrs, 0);
     }
 
-    public AffirmPromotionLabel(@NonNull Context context,
-                                @Nullable AttributeSet attrs,
-                                int defStyleAttr) {
+    public AffirmPromotionButton(@NonNull Context context,
+                                 @Nullable AttributeSet attrs,
+                                 int defStyleAttr) {
         super(context, attrs, defStyleAttr);
 
         init(context, attrs);
@@ -48,14 +48,14 @@ public class AffirmPromotionLabel extends AppCompatTextView {
 
     private void init(@NonNull Context context, @Nullable AttributeSet attrs) {
         TypedArray typedArray =
-                context.obtainStyledAttributes(attrs, R.styleable.AffirmPromotionLabel);
+                context.obtainStyledAttributes(attrs, R.styleable.AffirmPromotionButton);
 
         int affirmLogoTypeOrdinal =
-                typedArray.getInt(R.styleable.AffirmPromotionLabel_affirmLogoType,
+                typedArray.getInt(R.styleable.AffirmPromotionButton_affirmLogoType,
                         AFFIRM_DISPLAY_TYPE_TEXT.getOrdinal());
 
         int affirmColorOrdinal =
-                typedArray.getInt(R.styleable.AffirmPromotionLabel_affirmColor,
+                typedArray.getInt(R.styleable.AffirmPromotionButton_affirmColor,
                         AFFIRM_COLOR_TYPE_WHITE.getOrdinal());
 
         mAffirmLogoType = AffirmLogoType.getAffirmLogoType(affirmLogoTypeOrdinal);